+++ OUDS/Core/OUDS/Sources/OUDSTheme/OUDSTheme.swift
@@ -90,6 +90,9 @@ open class OUDSTheme: @unchecked Sendable {
     /// All components tokens related to chip components like `OUDSChip`
     public let chip: AllChipComponentTokensProvider
 
+    /// All components tokens related to breadcrumb components like `OUDSBreadcrumb`
+    public let breadcrumb: AllBreadcrumbComponentTokensProvider
+
     // NOTE: Add new component tokens provider here
 
     // MARK: - Initializers
@@ -129,6 +132,8 @@ open class OUDSTheme: @unchecked Sendable {
         `switch` = tokensProviders.get()
         listItem = tokensProviders.get()
         chip = tokensProviders.get()
+        breadcrumb = tokensProviders.get()
+
         // NOTE: Add new component tokens provider here
 
         // Load other configuration elements